Refactor : 변수 호출 수정
