+++ src/views/SignIn/BasicView.vue
@@ -143,6 +143,7 @@ onMounted(() => {
   </header>
 </template>
 <script>
+
 export default {
   name: "Login",
   data() {
@@ -153,8 +154,8 @@ export default {
   },
   methods: {
     loginWithKakao() {
-      const REST_API_KEY = "2e108c0efaed0b62af32afb27ff62354";
-      const REDIRECT_URI = "http://localhost:3000/oauth-redirect-kakao";
+      const REST_API_KEY = import.meta.env.VITE_APP_KAKAO_REST_API_KEY
+      const REDIRECT_URI = `${import.meta.env.VITE_APP_LOCAL_FRONTEND_URL}/oauth-redirect-kakao`;
       window.location.href =
         "https://kauth.kakao.com/oauth/authorize?response_type=code&client_id=" +
         REST_API_KEY +
@@ -162,16 +163,16 @@ export default {
         REDIRECT_URI;
     },
     loginWithNaver() {
-      const NAVER_CLIENT_ID = "tKBPdHduSdAwkRZZWFK2"
-      const NAVER_REDIRECT_URL = "http://localhost:3000/oauth-redirect-naver"
+      const NAVER_CLIENT_ID = import.meta.env.VITE_APP_NAVER_CLINET_ID
+      const NAVER_REDIRECT_URL = `${import.meta.env.VITE_APP_LOCAL_FRONTEND_URL}/oauth-redirect-naver`
       const url = 'https://nid.naver.com/oauth2.0/authorize?response_type=code&client_id=' + NAVER_CLIENT_ID + '&redirect_uri=' + NAVER_REDIRECT_URL;
 
       window.location.href = url;
     },
     loginWithGoogle() {
 
-      const GOOGLE_CLIENT_ID = "570712010140-1n043i16lahibthrj0ma9lqqj24mo0im.apps.googleusercontent.com"
-      const GOOGLE_REDIRECT_URL = "http://localhost:3000/oauth-redirect-google"
+      const GOOGLE_CLIENT_ID = import.meta.env.VITE_APP_GOOGLE_CLINET_ID
+      const GOOGLE_REDIRECT_URL = `${import.meta.env.VITE_APP_LOCAL_FRONTEND_URL}/oauth-redirect-google`
       const url = 'https://accounts.google.com/o/oauth2/v2/auth?client_id=' + GOOGLE_CLIENT_ID + '&redirect_uri=' + GOOGLE_REDIRECT_URL + '&response_type=code' + '&scope=email profile';
 
       window.location.href = url;
